Lagos Rent Prediction Web App
This project is an inter-track project that involves collaboration with the Software team, UI/UX team and the Data Science team for which I was a part of.  It involves building an application that will predict the rent prices for houses in Lagos State. We limited our scope to locations and number of rooms(1-3bedrooms). As part of the Data Science team, our role was to among others Scrape websites  and then build a Machine Learning model to help house seekers in Lagos have basic idea of how much the cost of renting an apartment would be,